chore(api): scaffold admin past-questions upload route
This commit introduces a new route for uploading past questions in the admin
section of the API. The addition is part of the ongoing effort to enhance
administrative functionalities. There are no immediate risks or side effects
as this is a scaffolded route without implemented logic. Further development
will be required to complete the functionality.